Set layout attributes and add strings
Strings were added to the strings file instead of being hard coded so
they can be changed later or could be written for different languages.

The layout attributes were used to set the style, color, etc to make it
more readable and follow along with the example. Does not follow the
example perfectly as this is also a personal experiment on how to draw
user interfaces.