Fix: Lower target and compile SDK version
SDK version 28 has a bug that was causing the application not to render
within Android studio.  When the application is pushed to a real device
the application would crash.  The solution was to lower to a stable
version of the SDK (version 26) and compile. This corrected the render
issue with in Android Studio.