Fix: Lower target and compile SDK version
SDK version 28 has a bug that was causing the application not to render
within Android Studio.  When the application is pushed to a real device
the application would crash. The solution was to lower to a stable
version of the SDK (version 26) and compile. This corrected the render
issue within in Android Studio and physical devices.

     Cherrypick commit:
        6cf59cacbc1dd6fd37c45460570950bc4eb456ef